requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
datasets>=2.14.0
evaluate>=0.4.0
accelerate>=0.24.0

openai-clip-torch>=2.0.2
sentencepiece
timm
opencv-python
Pillow>=9.0

torchattacks>=3.4.0
torchopt>=0.5.0

scikit-learn>=1.3.0
nltk>=3.8.1
bert-score>=0.3.13
pycocoevalcap @ git+https://github.com/salaniz/pycocoevalcap.git

jaxtyping>=0.2.19
einops>=0.6.0
requests>=2.31.0
PyYAML>=6.0
tqdm>=4.66.0
matplotlib>=3.7.0
Levenshtein>=0.21.0
huggingface-hub>=0.18.0
